Improve public API
- Rename classes
- Improving CLI usage
